+++ Sources/StrasbourgParkAPI/ParkingAPIClient.swift
@@ -14,6 +14,8 @@ public enum ParkingAPIClientError: Error {
     case network(Error)
     /// An error due to decoding error
     case decodable(Error)
+    /// An unknown error type
+    case generic(Error)
 }
 
 /// An http client to query
@@ -28,7 +30,6 @@ public final class ParkingAPIClient: NSObject, URLSessionDelegate {
     /// - Parameter configuration: The `URLSessionConfiguration` to use. Default to `URLSession.default`
     /// - Parameter pageSize: The pagination value to use. Default to `10`.
     public init(configuration: URLSessionConfiguration = .default, pageSize: UInt = 10) {
-
         let queue = OperationQueue()
         queue.qualityOfService = .background
         queue.name = "net.yageek.strasbourgpark.apiclient"
@@ -58,7 +59,7 @@ public final class ParkingAPIClient: NSObject, URLSessionDelegate {
 
     // MARK: - Legacy APIs
 
-    /// Retrieve the parkings' locations with the legacy endpoints
+    /// Retrieve the parkings' locations with the legacy API
     /// - Parameter completion: The result when the request completed
     /// - Returns: A `CancelableRequest` compatible element
     @discardableResult public func getLegacyLocation(completion: @escaping(Result<LocationResponse, ParkingAPIClientError>) -> Void) -> CancelableRequest {
@@ -67,7 +68,7 @@ public final class ParkingAPIClient: NSObject, URLSessionDelegate {
         return op
     }
 
-    /// Retreive the parkings' with the legacy API
+    /// Retrieve the parkings' statuses with the legacy API
     /// - Parameter completion: The result when the request completed
     /// - Returns: A `CancelableRequest` compatible element
    @discardableResult public func getLegacyStatus(completion: @escaping(Result<StatusResponse, ParkingAPIClientError>) -> Void) -> CancelableRequest {
@@ -77,20 +78,20 @@ public final class ParkingAPIClient: NSObject, URLSessionDelegate {
     }
 
     // MARK: - Open Data APIS
-    /// Retrieve the parkings' locations
+    /// Retrieve the parkings' locations with the open data API
     /// - Parameter completion: The result when the request completed
     /// - Returns: A `CancelableRequest` compatible element
-    @discardableResult public func getLocations(completion: @escaping(Result<[LocationOpenData], Error>) -> Void) -> CancelableRequest {
+    @discardableResult public func getLocations(completion: @escaping(Result<[LocationOpenData], ParkingAPIClientError>) -> Void) -> CancelableRequest {
         let op = DownloadAllPages<LocationOpenData>(session: self.session, endpoint: .location, pageSize: self.pageSize, completionHandler: completion)
         workingQueue.addOperation(op)
         return op
     }
 
     // MARK: - Open Data APIS
-    /// Retreive the parkings' with the legacy API
+    /// Retrieve the parkings' status with the open data API
     /// - Parameter completion: The result when the request completed
     /// - Returns: A `CancelableRequest` compatible element
-    @discardableResult public func getStatus(completion: @escaping(Result<[StatusOpenData], Error>) -> Void) -> CancelableRequest {
+    @discardableResult public func getStatus(completion: @escaping(Result<[StatusOpenData], ParkingAPIClientError>) -> Void) -> CancelableRequest {
         let op = DownloadAllPages<StatusOpenData>(session: self.session, endpoint: .status, pageSize: self.pageSize, completionHandler: completion)
         workingQueue.addOperation(op)
         return op
@@ -169,34 +170,32 @@ extension Publishers {
 @available(iOS 13.0, macOS 10.15, *)
 extension ParkingAPIClient {
     
-    /// Retrieve the parkings' locations with the legacy endpoints
-    /// - Returns: A ``AnyPublisher`` instance providing legacy ``LocationResponse`` element.
+    /// Retrieve the parkings' locations with the legacy API
+    /// - Returns: A ``AnyPublisher`` answering one ``LocationResponse`` element.
     public func getLegacyLocationPublisher() -> AnyPublisher<LocationResponse, ParkingAPIClientError> {
         let op = DownloadOperation<LocationResponse>(session: self.session, endpoint: .legacyLocation)
         return Publishers.ParkingPublisher(operation: op, queue: self.workingQueue).eraseToAnyPublisher()
     }
 
-    /// Retreive the parkings' statuses with the legacy API
-    /// - Returns: A ``AnyPublisher`` instance providing legacy ``StatusResponse`` element.
+    /// Retrieve the parkings' statuses with the legacy API
+    /// - Returns: A ``AnyPublisher`` answering one ``StatusResponse`` element.
     public func getLegacyStatusPublisher() -> AnyPublisher<StatusResponse, ParkingAPIClientError> {
         let op = DownloadOperation<StatusResponse>(session: self.session, endpoint: .legacyLocation)
         return Publishers.ParkingPublisher(operation: op, queue: self.workingQueue).eraseToAnyPublisher()
     }
 
     // MARK: - Open Data APIS
-    /// Retrieve the parkings' locations
-    /// - Parameter completion: The result when the request completed
-    /// - Returns: A ``AnyPublisher`` instance providing legacy ``[LocationOpenData]`` element.
-    public func getLocationsPublisher() -> AnyPublisher<[LocationOpenData], Error> {
+    /// Retrieve the parkings' locations using the open data API
+    /// - Returns: A ``AnyPublisher`` answering one ``[LocationOpenData]`` .
+    public func getLocationsPublisher() -> AnyPublisher<[LocationOpenData], ParkingAPIClientError> {
         let op = DownloadAllPages<LocationOpenData>(session: self.session, endpoint: .location, pageSize: self.pageSize)
         return Publishers.ParkingPublisher(operation: op, queue: self.workingQueue).eraseToAnyPublisher()
     }
 
     // MARK: - Open Data APIS
-    /// Retreive the parkings' with the legacy API
-    /// - Parameter completion: The result when the request completed
-    /// - Returns: A ``AnyPublisher`` instance providing legacy ``[StatusOpenData]`` element.
-    public func getStatusPublisher() -> AnyPublisher<[StatusOpenData], Error> {
+    /// Retrieve the parkings' statuses with the opendata API
+    /// - Returns: A ``AnyPublisher`` answering one ``[StatusOpenData]`` element.
+    public func getStatusPublisher() -> AnyPublisher<[StatusOpenData], ParkingAPIClientError> {
         let op = DownloadAllPages<StatusOpenData>(session: self.session, endpoint: .status, pageSize: self.pageSize)
         return Publishers.ParkingPublisher(operation: op, queue: self.workingQueue).eraseToAnyPublisher()
     }
@@ -246,28 +245,32 @@ extension ParkingAPIClient {
         }
     }
 
-
-    /// Retrieve the legacy locations
+    /// Retrieve the parkings' location with the legacy API
+    /// - Returns: An array of ``LocationResponse``
     public func fetchLegacyLocation() async throws -> LocationResponse {
         let dlOp = DownloadOperation<LocationResponse>(session: self.session, endpoint: .legacyLocation)
         return try await self.executeOperation(dlOp)
     }
-
+    /// Retrieve the parkings' statuses with the legacy API
+    /// - Returns: An array of ``StatusResponse``
     public func fetchLegacyStatus() async throws -> StatusResponse {
         let op = DownloadOperation<StatusResponse>(session: self.session, endpoint: .legacyLocation)
         return try await self.executeOperation(op)
     }
 
     // MARK: - Open Data APIS
+
+    /// Retrieve the parkings' location with the open data API
+    /// - Returns: An array of ``LocationOpenData``
+    public func fetchLocations() async throws -> [LocationOpenData] {
+        let dlOp = DownloadAllPages<LocationOpenData>(session: self.session, endpoint: .location, pageSize: self.pageSize)
+        return try await self.executeOperation(dlOp)
+    }
+
+    /// Retrieve the parkings' statuses with the open data API
+    /// - Returns: An array of ``StatusResponse``
     public func fetchStatus() async throws -> [StatusOpenData] {
         let op = DownloadAllPages<StatusOpenData>(session: self.session, endpoint: .status, pageSize: self.pageSize)
         return try await self.executeOperation(op)
-    }
-
-    /// Retrieve all the locations of parkings
-    /// - Returns: An array of `StrasbourgParkAPI.LocationOpenData`
-    public func fetchLocations() async throws -> [StrasbourgParkAPI.LocationOpenData] {
-        let dlOp = DownloadAllPages<LocationOpenData>(session: self.session, endpoint: .location, pageSize: self.pageSize)
-        return try await self.executeOperation(dlOp)
     }
 }